refactor: extract request routing into SRP core microcrate
