+++ Cargo.toml
@@ -61,6 +61,7 @@ members = [
   "crates/bitnet-api-key-auth-core", # SRP: API key auth policy/store primitives
   "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
   "crates/bitnet-request-context-core", # SRP: reusable request context metadata contracts
+  "crates/bitnet-request-router-core", # SRP: reusable HTTP route matching + router tables
   "crates/bitnet-server",
   "crates/bitnet-cli-sampling-core", # SRP: reusable CLI sampling primitives
   "crates/bitnet-cli-config-core", # SRP: reusable CLI configuration contracts/validation
@@ -181,6 +182,7 @@ default-members = [
   "crates/bitnet-feature-matrix",
   "crates/bitnet-server-health-types-core",
   "crates/bitnet-request-context-core",
+  "crates/bitnet-request-router-core",
   "crates/bitnet-testing-policy-core",
   "crates/bitnet-testing-policy-snapshot-core", # SRP: policy snapshot resolution helpers
   "crates/bitnet-testing-scenarios-core",
